Upgraded proto related deps: grpc and protobuf; removed gogo from core.
