+++ .bingo/protoc-gen-go.mod
@@ -2,4 +2,4 @@ module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
 
 go 1.14
 
-require github.com/bwplotka/bingo v0.2.3
+require google.golang.org/protobuf v1.25.0 // cmd/protoc-gen-go
