+++ .bingo/variables.env
@@ -8,7 +8,7 @@ if [ -z "$gobin" ]; then
 fi
 
 
-BINGO="${gobin}/bingo-v0.2.2"
+BINGO="${gobin}/bingo-v0.2.3"
 
 FAILLINT="${gobin}/faillint-v1.5.0"
 
@@ -17,6 +17,10 @@ GOIMPORTS="${gobin}/goimports-v0.0.0-20200529172331-a64b76657301"
 GOLANGCI_LINT="${gobin}/golangci-lint-v1.27.0"
 
 MISSPELL="${gobin}/misspell-v0.3.4"
+
+PROTOC_GEN_GO_GRPC="${gobin}/protoc-gen-go-grpc-v0.0.0-20200723182653-9106c3fff523"
+
+PROTOC_GEN_GO="${gobin}/protoc-gen-go-v1.25.0"
 
 PROTOC_GEN_GOGOFAST="${gobin}/protoc-gen-gogofast-v1.3.1"
