+++ go.mod
@@ -3,13 +3,13 @@ module github.com/grpc-ecosystem/go-grpc-middleware/v2
 go 1.14
 
 require (
-	github.com/gogo/protobuf v1.2.1
-	github.com/golang/protobuf v1.3.2
+	github.com/golang/protobuf v1.4.2
 	github.com/opentracing/opentracing-go v1.1.0
 	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.4.0
+	github.com/stretchr/testify v1.6.1
 	golang.org/x/net v0.0.0-20190620200207-3b0461eec859
 	golang.org/x/oauth2 v0.0.0-20180821212333-d2e6202438be
-	google.golang.org/genproto v0.0.0-20190819201941-24fa4b261c55 // indirect
-	google.golang.org/grpc v1.19.0
+	google.golang.org/grpc v1.30.0
+	google.golang.org/grpc/examples v0.0.0-20200723182653-9106c3fff523
+	google.golang.org/protobuf v1.25.0
 )
