+++ grpctesting/external/gogofieldstestpb/fields.proto
@@ -1,14 +1,14 @@
 syntax = "proto3";
 
-//  This file is used for grpctesting discovery of log fields from requests using reflection and gogo proto more tags.
-package grpc_middleware.gogotestpb;
+//  This file is used for grpctesting/external discovery of log fields from requests using reflection and gogo proto more tags.
+package grpc_middleware.gogofieldstestpb;
 
 import "gogoproto/gogo.proto";
 import "google/protobuf/timestamp.proto";
 
 option (gogoproto.gogoproto_import) = false;
 
-option go_package = "gogotestpb";
+option go_package = "gogofieldstestpb";
 
 message Metadata {
     repeated string tags = 1 [(gogoproto.moretags) = "log_field:\"meta_tags\""];
