+++ providers/zap/go.mod
@@ -5,5 +5,5 @@ go 1.14
 require (
 	github.com/grpc-ecosystem/go-grpc-middleware/v2 v2.0.0-20200501113911-9a95f0fdbfea
 	go.uber.org/zap v1.15.0
-	google.golang.org/grpc v1.19.0
+	google.golang.org/grpc v1.30.0
 )
