+++ providers/zerolog/go.mod
@@ -5,5 +5,5 @@ go 1.14
 require (
 	github.com/grpc-ecosystem/go-grpc-middleware/v2 v2.0.0-20200501113911-9a95f0fdbfea
 	github.com/rs/zerolog v1.19.0
-	google.golang.org/grpc v1.19.0
+	google.golang.org/grpc v1.30.0
 )
